refactor search page
