+++ package.json
@@ -8,7 +8,9 @@
     "@emotion/styled": "^11.10.4",
     "@mui/icons-material": "^5.10.6",
     "@mui/material": "^5.10.6",
-    "@polkadot/keyring": "^10.1.8",
+    "@polkadot/keyring": "^10.1.12",
+    "@polkadot/util": "^10.1.12",
+    "@polkadot/util-crypto": "^10.1.12",
     "@rollbar/react": "^0.11.1",
     "@subsquid/archive-registry": "^1.0.14",
     "@testing-library/jest-dom": "^5.16.5",
