+++ src/components/Loading.tsx
@@ -4,7 +4,6 @@ import { css } from "@emotion/react";
 import Spinner from "./Spinner";
 
 const loadingStyle = css`
-	padding: 16px 0;
 	text-align: center;
 `;
