+++ test/e2e/searchResults.spec.ts
@@ -2,15 +2,16 @@ import { test, expect } from "@playwright/test";
 
 import { removeContent } from "../utils/removeContent";
 import { screenshot } from "../utils/screenshot";
+import { waitForPageEvent } from "../utils/waitForPageEvent";
 
 test.describe("Search results page", () => {
-	test.beforeEach(async ({ page }) => {
-		await page.goto("/equilibrium/search?query=Oracle.NewPrice");
-		await page.waitForTimeout(2000); // TODO find out better waiting until page is loaded
-	});
-
 	test("shows search results page", async ({ page }) => {
+		await page.goto("/equilibrium/search?query=Oracle.NewPrice");
+		await waitForPageEvent(page, "data-loaded");
+
 		await removeContent(page.locator("[data-test=events-table] tr td"));
 		await screenshot(page, "searchResults");
 	});
+
+
 });
